use __uintptr_type__
git-svn-id: file:///raid/svn-personal/tewi/trunk@31 8739d7e6-ffea-ec47-b151-bdff447c6205

+++ Server/server.c
@@ -502,7 +502,7 @@ void tw_server_pass(int sock, bool ssl, int port, SOCKADDR addr) {
 										continue;
 									}
 									if(s.st_size < 1024ULL) {
-										sprintf(size, "%d", s.st_size);
+										sprintf(size, "%d", (int)s.st_size);
 									} else if(s.st_size < 1024ULL * 1024) {
 										sprintf(size, "%.1fK", (double)s.st_size / 1024);
 									} else if(s.st_size < 1024ULL * 1024 * 1024) {
